// src/app/models/accommodations/accommodations.model.ts
export class AccommodationsDTO {
  id: string | undefined | null = null;
  createdAt: Date | undefined | null = null;
  priceInEur: string | undefined | null = null;
  avaliable: boolean | undefined | null = null;
  city: string | undefined | null = null;
  address: string | undefined | null = null;
  country: string | undefined | null = null;
  name: string | undefined | null = null;
}